Split up file for each slice of reducer
