+++ 10-diving-redux/src/store/authSlice.js
@@ -1,0 +1,21 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+const initialAuthState = {
+    isAuthenticated: false,
+};
+
+const authSlice = createSlice({
+    name: 'Auth',
+    initialState: initialAuthState,
+    reducers: {
+        login(state) {
+            state.isAuthenticated = true;
+        },
+        logout(state) {
+            state.isAuthenticated = false;
+        },
+    },
+});
+
+export const authAction = authSlice.actions;
+export default authSlice.reducer
