+++ 10-diving-redux/src/store/index.js
@@ -1,48 +1,12 @@
-import { configureStore, createSlice } from '@reduxjs/toolkit';
+import { configureStore } from '@reduxjs/toolkit';
+import authSlice from './authSlice';
+import counterSlice from './counterSlice';
 
-const initialCounterState = { counter: 0, showCounter: true };
-const counterSlice = createSlice({
-    name: 'counter',
-    initialState: initialCounterState,
-    reducers: {
-        increment(state) {
-            state.counter++;
-        },
-        decrement(state) {
-            state.counter--;
-        },
-        increase(state, action) {
-            state.counter = state.counter + action.payload;
-        },
-        toggleCounter(state) {
-            state.showCounter = !state.showCounter;
-        },
-    },
-});
-
-const initialAuthState = {
-    isAuthenticated: false,
-};
-
-const authSlice = createSlice({
-    name: 'Auth',
-    initialState: initialAuthState,
-    reducers: {
-        login(state) {
-            state.isAuthenticated = true;
-        },
-        logout(state) {
-            state.isAuthenticated = false;
-        },
-    },
-});
 const store = configureStore({
     reducer: {
-        counter: counterSlice.reducer,
-        auth: authSlice.reducer,
+        counter: counterSlice,
+        auth: authSlice,
     },
 });
 
-export const counterActions = counterSlice.actions;
-export const authAction = authSlice.actions;
 export default store;
